# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Swagger Gentcmen API documentation",
        "contact": {
            "email": "nikitosnov@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Fetch users by query params",
                "operationId": "Fetch users by query params",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_admin",
                        "in": "query",
                        "description": "Fetch admins",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "user_city",
                        "in": "query",
                        "description": "User's city",
                        "schema": {
                            "type": "string"
                        },
                        "example": "some user name"
                    },
                    {
                        "name": "user_name",
                        "in": "query",
                        "description": "User's name",
                        "schema": {
                            "type": "string"
                        },
                        "example": "some user name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users fetched successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/users/{user}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Fetch user by id",
                "operationId": "Fetch user by id",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User fetched successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "404": {
                        "description": "Entity not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Remove user",
                "operationId": "Remove user",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User was deleted successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/admin/users/{user}/coupons": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Fetch user's coupons",
                "operationId": "Fetch user's coupons",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User's coupons fetched successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "404": {
                        "description": "Entity not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/users/create-admin": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create a new admin",
                "operationId": "Create admin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAdminVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Admin created successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/{admin}": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update admin by id",
                "operationId": "Update admin by id",
                "parameters": [
                    {
                        "name": "admin",
                        "in": "path",
                        "description": "Admin id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchUpdateAdminVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admin is updated successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "404": {
                        "description": "Admin not found with provided id"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/users/{user}/manage-bonus-points": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Manage bonus points",
                "operationId": "manage bonus points",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "title": "Amount",
                                        "description": "Status of order",
                                        "example": 200
                                    },
                                    "operation": {
                                        "title": "Operation",
                                        "description": "increment or decrement",
                                        "example": "increment"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bonus points updated successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "404": {
                        "description": "Entity not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/admin/users/{user}/manage-coupons": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Manage coupons",
                "operationId": "manage coupons",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DestroySomeItemsVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Coupons for user updated successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "404": {
                        "description": "Entity not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Remove user's coupon",
                "operationId": "Remove user's coupon",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ids": {
                                        "title": "ids",
                                        "description": "Identifiers that should be deleted",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    },
                                    "forceDelete": {
                                        "title": "forceDelete",
                                        "description": "Identifiers that should be deleted anyway",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Coupons for user updated successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateAdminVirtualBody": {
                "title": "Store Admin request",
                "description": "Store Admin request body data",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new user",
                        "type": "string",
                        "example": "test_user"
                    },
                    "email": {
                        "title": "email",
                        "description": "Email of the new user",
                        "type": "string",
                        "example": "test@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "Password of the new user",
                        "type": "string",
                        "example": "somepassword"
                    }
                },
                "type": "object"
            },
...
            "DestroySomeItemsVirtualBody": {
                "title": "Destroy some items example",
                "description": "Destroy some items body data",
                "required": [
                    "ids"
                ],
                "properties": {
                    "ids": {
                        "title": "ids",
                        "description": "Identifiers that should be deleted",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2,
                            3
                        ]
                    }
                },
                "type": "object"
            },
...
            "PatchUpdateAdminVirtualBody": {
                "title": "Update Admin request",
                "description": "Update Admin request body data",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new user",
                        "type": "string",
                        "example": "test_user"
                    },
                    "email": {
                        "title": "email",
                        "description": "Email of the new user",
                        "type": "string",
                        "example": "test@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "Password of the new user",
                        "type": "string",
                        "example": "somepassword"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}